resolvers updated, graphql error, and automatic nav for specific condition
