+++ src/app/graphql/resolvers/mutations/products.ts
@@ -49,7 +49,7 @@ export const productMutation = {
                             });
                         }
 
-                        throw new GraphQLError("Database is unreachable.", {
+                        throw new GraphQLError("Database is unreachable. Make sure your database are running.", {
                             extensions: {
                             code: "NO_DATABASE_FOUND"
                             },
@@ -102,7 +102,7 @@ export const productMutation = {
                                 });
                             }
 
-                            throw new GraphQLError("Database is unreachable.", {
+                            throw new GraphQLError("Database is unreachable. Make sure your database are running.", {
                                 extensions: {
                                 code: "NO_DATABASE_FOUND"
                                 },
@@ -201,7 +201,7 @@ export const productMutation = {
                                     },
                                 });
                             }
-                            throw new GraphQLError("Database is unreachable.", {
+                            throw new GraphQLError("Database is unreachable. Make sure your database are running.", {
                                 extensions: {
                                 code: "NO_DATABASE_FOUND"
                                 },
