+++ src/app/graphql/resolvers/mutations/users.ts
@@ -51,7 +51,7 @@ export const usersMutation = {
                             });
                         }
 
-                        throw new GraphQLError("Database is unreachable.", {
+                        throw new GraphQLError("Database is unreachable. Make sure your database are running.", {
                             extensions: {
                                code: "NO_DATABASE_FOUND"
                             },
